fix: move options to custom file
